server/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ML and Vector Search
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2

# AI/LLM Integration
google-generativeai==0.3.2

# Data Processing
pyarrow==14.0.1
sentence-transformers==2.7.0
huggingface_hub==0.19.4

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Logging and Monitoring
structlog==23.2.0
rich==13.7.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production
gunicorn==21.2.0